*: resolve check dependency.
